--- Game2/core/src/com/haruham/game/level/Event.java
@@ -1,7 +1,0 @@
-package com.haruham.game.level;
-
-/**
- * Created on 10/1/2015.
- */
-public class Event {
-}
