+++ Game2/core/src/com/haruham/game/util/Event.java
@@ -1,0 +1,7 @@
+package com.haruham.game.util;
+
+/**
+ * Created on 10/4/2015.
+ */
+public class Event {
+}
